+++ packages/core/src/controllers/BlockchainApiController.ts
@@ -10,11 +10,10 @@ const api = new FetchUtil({ baseUrl: 'https://rpc.walletconnect.com' });
 
 // -- Controller ---------------------------------------- //
 export const BlockchainApiController = {
-  fetchIdentity({ caipChainId, address }: BlockchainApiIdentityRequest) {
+  fetchIdentity({ address }: BlockchainApiIdentityRequest) {
     return api.get<BlockchainApiIdentityResponse>({
       path: `/v1/identity/${address}`,
       params: {
-        chainId: caipChainId,
         projectId: OptionsController.state.projectId
       }
     });