+++ packages/email-wagmi/src/index.ts
@@ -1,15 +1,7 @@
-import { type Chain, type ConnectorData, Connector } from 'wagmi';
-import { SwitchChainError, createWalletClient, custom, getAddress } from 'viem';
+import { createConnector, ChainNotConfiguredError } from 'wagmi';
+import { SwitchChainError, getAddress, type Address } from 'viem';
+import type { Evaluate } from '@wagmi/core/internal';
 import { W3mFrameProvider } from '@web3modal/email-react-native';
-
-export type StorageStoreData = {
-  state: { data?: ConnectorData };
-};
-
-interface Config {
-  chains?: Chain[];
-  options: EmailProviderOptions;
-}
 
 export type Metadata = {
   name: string;
@@ -18,129 +10,124 @@ export type Metadata = {
   icons: string[];
 };
 
-type EmailProviderOptions = {
+type EmailProviderOptions = Evaluate<{
   /**
    * WalletConnect Cloud Project ID.
    * @link https://cloud.walletconnect.com/sign-in.
    */
   projectId: string;
   metadata: Metadata;
+}>;
+
+type Provider = W3mFrameProvider;
+
+type StorageItemMap = {
+  '@w3m/connected_connector'?: string;
 };
 
-export class EmailConnector extends Connector<W3mFrameProvider, EmailProviderOptions> {
-  readonly id = 'w3mEmail';
-  readonly name = 'Web3Modal Email';
-  readonly ready = true;
+emailConnector.type = 'w3mEmail' as const;
+emailConnector.id = 'w3mEmail' as const;
+export function emailConnector(parameters: EmailProviderOptions) {
+  let _provider: W3mFrameProvider = {} as W3mFrameProvider;
 
-  private provider: W3mFrameProvider = {} as W3mFrameProvider;
-
-  constructor(config: Config) {
-    super(config);
-    this.provider = new W3mFrameProvider(config.options.projectId, config.options.metadata);
-  }
-
-  async connect(options: { chainId?: number }): Promise<Required<ConnectorData>> {
-    const provider = await this.getProvider();
-    await provider.webviewLoadPromise;
-
-    const { address, chainId } = await provider.connect({ chainId: options?.chainId });
-
-    return {
-      account: address as `0x${string}`,
-      chain: {
-        id: chainId,
-        unsupported: this.isChainUnsupported(1)
-      }
-    };
-  }
-
-  async disconnect(): Promise<void> {
-    const provider = await this.getProvider();
-    await provider.webviewLoadPromise;
-
-    await provider.disconnect();
-  }
-
-  override async switchChain(chainId: number): Promise<Chain> {
-    try {
-      const chain = this.chains?.find(c => c.id === chainId);
-      if (!chain) {
-        throw new SwitchChainError(new Error('chain not found on connector.'));
-      }
-
+  return createConnector<Provider, {}, StorageItemMap>(config => ({
+    id: emailConnector.id,
+    name: 'Web3Modal Email',
+    type: emailConnector.type,
+    async setup() {
+      _provider = new W3mFrameProvider(parameters.projectId, parameters.metadata);
+    },
+    async connect(options = {}) {
       const provider = await this.getProvider();
       await provider.webviewLoadPromise;
-      await provider.switchNetwork(chainId);
+      const { address, chainId } = await provider.connect({ chainId: options.chainId });
 
-      const unsupported = this.isChainUnsupported(chainId);
-      this.emit('change', { chain: { id: chainId, unsupported } });
+      return {
+        accounts: [address as Address],
+        account: address as Address,
+        chainId,
+        chain: {
+          id: chainId,
+          unsuported: false
+        }
+      };
+    },
+    async disconnect() {
+      const provider = await this.getProvider();
+      await provider.webviewLoadPromise;
+      await provider.disconnect();
+    },
+    async switchChain({ chainId }) {
+      try {
+        const chain = config.chains?.find(c => c.id === chainId);
+        if (!chain) throw new SwitchChainError(new ChainNotConfiguredError());
 
-      return chain;
-    } catch (error) {
-      if (error instanceof Error) {
-        throw new SwitchChainError(error);
+        const provider = await this.getProvider();
+        await provider.webviewLoadPromise;
+        await provider.switchNetwork(chainId);
+        config.emitter.emit('change', { chainId: Number(chainId) });
+
+        return chain;
+      } catch (error) {
+        if (error instanceof Error) {
+          throw new SwitchChainError(error);
+        }
+        throw error;
       }
-      throw error;
+    },
+    async getAccounts() {
+      const provider = await this.getProvider();
+      await provider.webviewLoadPromise;
+
+      return (
+        await provider.request({
+          method: 'eth_accounts'
+        })
+      ).map(getAddress);
+    },
+    async getChainId() {
+      const provider = await this.getProvider();
+      await provider.webviewLoadPromise;
+      const { chainId } = await provider.getChainId();
+
+      return chainId;
+    },
+    async getProvider() {
+      return Promise.resolve(_provider);
+    },
+    async isAuthorized() {
+      try {
+        const provider = await this.getProvider();
+        await provider.webviewLoadPromise;
+        const connectedConnector = await config.storage?.getItem('recentConnectorId');
+
+        if (connectedConnector !== emailConnector.id) {
+          // isConnected still needs to be called to disable email input loader
+          provider.isConnected();
+
+          return false;
+        } else {
+          const { isConnected } = await provider.isConnected();
+
+          return isConnected;
+        }
+      } catch (error) {
+        return false;
+      }
+    },
+    onAccountsChanged(accounts) {
+      if (accounts.length === 0) config.emitter.emit('disconnect');
+      else config.emitter.emit('change', { accounts: accounts.map(getAddress) });
+    },
+    onChainChanged(chain) {
+      const chainId = Number(chain);
+      config.emitter.emit('change', { chainId });
+    },
+    async onDisconnect() {
+      const provider = await this.getProvider();
+      await provider.webviewLoadPromise;
+      await provider.disconnect();
+      config.emitter.emit('disconnect');
     }
-  }
-
-  async getAccount(): Promise<`0x${string}`> {
-    const provider = await this.getProvider();
-    await provider.webviewLoadPromise;
-    const { address } = await provider.connect();
-
-    return address as `0x${string}`;
-  }
-
-  async getChainId(): Promise<number> {
-    const provider = await this.getProvider();
-    await provider.webviewLoadPromise;
-    const { chainId } = await provider.getChainId();
-
-    return chainId;
-  }
-
-  async getProvider() {
-    return Promise.resolve(this.provider);
-  }
-
-  async getWalletClient() {
-    const provider = await this.getProvider();
-    await provider.webviewLoadPromise;
-    const { address, chainId } = await provider.connect();
-
-    return Promise.resolve(
-      createWalletClient({
-        account: address as `0x${string}`,
-        chain: { id: chainId } as Chain,
-        transport: custom(this.provider)
-      })
-    );
-  }
-
-  async isAuthorized(): Promise<boolean> {
-    const provider = await this.getProvider();
-    await provider.webviewLoadPromise;
-    const { isConnected } = await provider.isConnected();
-
-    return isConnected;
-  }
-
-  protected onAccountsChanged = (accounts: string[]) => {
-    if (accounts.length === 0) this.emit('disconnect');
-    else this.emit('change', { account: getAddress(accounts[0]!) });
-  };
-
-  protected onChainChanged = (chainId: number | string) => {
-    const id = Number(chainId);
-    const unsupported = this.isChainUnsupported(id);
-    this.emit('change', { chain: { id, unsupported } });
-  };
-
-  async onDisconnect() {
-    const provider = await this.getProvider();
-    await provider.webviewLoadPromise;
-    await provider.disconnect();
-    this.emit('disconnect');
-  }
+  }));
 }