+++ packages/wagmi/src/client.ts
@@ -1,20 +1,17 @@
+import { formatUnits, type Hex } from 'viem';
 import {
-  type Address,
-  type Chain,
-  type Config,
-  type Connector as WagmiConnector,
+  type GetAccountReturnType,
   connect,
   disconnect,
-  fetchBalance,
-  fetchEnsAvatar,
-  fetchEnsName,
-  getAccount,
-  getNetwork,
-  switchNetwork,
+  switchChain,
   watchAccount,
-  watchNetwork
+  watchConnectors,
+  getEnsName,
+  getEnsAvatar as wagmiGetEnsAvatar,
+  getBalance
 } from '@wagmi/core';
-import { mainnet } from '@wagmi/core/chains';
+import { mainnet, type Chain } from '@wagmi/core/chains';
+import { EthereumProvider } from '@walletconnect/ethereum-provider';
 import {
   type CaipAddress,
   type CaipNetwork,
@@ -38,15 +35,13 @@ import {
   getEmailCaipNetworks,
   getWalletConnectCaipNetworks
 } from './utils/helpers';
+import { defaultWagmiConfig } from './utils/defaultWagmiConfig';
 
 // -- Types ---------------------------------------------------------------------
-interface WagmiConfig extends Config<any, any> {
-  connectors: WagmiConnector<any, any>[];
-}
+type WagmiConfig = ReturnType<typeof defaultWagmiConfig>;
 
 export interface Web3ModalClientOptions extends Omit<LibraryOptions, 'defaultChain' | 'tokens'> {
   wagmiConfig: WagmiConfig;
-  chains?: Chain[];
   defaultChain?: Chain;
   chainImages?: Record<number, string>;
   connectorImages?: Record<string, string>;
@@ -66,8 +61,10 @@ export class Web3Modal extends Web3ModalScaffold {
 
   private options: Web3ModalClientOptions | undefined = undefined;
 
+  private wagmiConfig: WagmiConfig;
+
   public constructor(options: Web3ModalClientOptions) {
-    const { wagmiConfig, chains, defaultChain, tokens, _sdkVersion, ...w3mOptions } = options;
+    const { wagmiConfig, defaultChain, tokens, _sdkVersion, ...w3mOptions } = options;
 
     if (!wagmiConfig) {
       throw new Error('web3modal:constructor - wagmiConfig is undefined');
@@ -81,7 +78,7 @@ export class Web3Modal extends Web3ModalScaffold {
       switchCaipNetwork: async caipNetwork => {
         const chainId = HelpersUtil.caipNetworkIdToNumber(caipNetwork?.id);
         if (chainId) {
-          await switchNetwork({ chainId });
+          await switchChain(wagmiConfig, { chainId });
         }
       },
 
@@ -117,15 +114,16 @@ export class Web3Modal extends Web3ModalScaffold {
           );
         }
 
-        connector.on('message', event => {
-          if (event.type === 'display_uri') {
-            onUri(event.data as string);
-            connector.removeAllListeners();
-          }
+        const provider = (await connector.getProvider()) as Awaited<
+          ReturnType<(typeof EthereumProvider)['init']>
+        >;
+
+        provider.on('display_uri', data => {
+          onUri(data);
         });
 
         const chainId = HelpersUtil.caipNetworkIdToNumber(this.getCaipNetwork()?.id);
-        await connect({ connector, chainId });
+        await connect(this.wagmiConfig, { connector, chainId });
       },
 
       connectExternal: async ({ id }) => {
@@ -135,10 +133,12 @@ export class Web3Modal extends Web3ModalScaffold {
         }
 
         const chainId = HelpersUtil.caipNetworkIdToNumber(this.getCaipNetwork()?.id);
-        await connect({ connector, chainId });
+        await connect(this.wagmiConfig, { connector, chainId });
       },
 
-      disconnect
+      disconnect: async () => {
+        await disconnect(this.wagmiConfig);
+      }
     };
 
     super({
@@ -151,15 +151,21 @@ export class Web3Modal extends Web3ModalScaffold {
     });
 
     this.options = options;
+    this.wagmiConfig = wagmiConfig;
 
-    this.syncRequestedNetworks(chains);
+    this.syncRequestedNetworks([...wagmiConfig.chains]);
+    this.syncConnectors([...wagmiConfig.connectors]);
+    this.listenEmailConnector([...wagmiConfig.connectors]);
 
-    this.syncConnectors(wagmiConfig);
-    this.syncEmailConnector(wagmiConfig);
-    this.listenEmailConnector(wagmiConfig);
+    watchConnectors(wagmiConfig, {
+      onChange: connectors => this.syncConnectors([...connectors])
+    });
 
-    watchAccount(() => this.syncAccount());
-    watchNetwork(() => this.syncNetwork());
+    watchAccount(wagmiConfig, {
+      onChange: accountData => {
+        this.syncAccount({ ...accountData });
+      }
+    });
   }
 
   // -- Public ------------------------------------------------------------------
@@ -185,7 +191,7 @@ export class Web3Modal extends Web3ModalScaffold {
   }
 
   // -- Private -----------------------------------------------------------------
-  private syncRequestedNetworks(chains: Web3ModalClientOptions['chains']) {
+  private syncRequestedNetworks(chains: Chain[]) {
     const requestedCaipNetworks = chains?.map(
       chain =>
         ({
@@ -198,17 +204,29 @@ export class Web3Modal extends Web3ModalScaffold {
     this.setRequestedCaipNetworks(requestedCaipNetworks ?? []);
   }
 
-  private async syncAccount() {
-    const { address, isConnected } = getAccount();
-    const { chain } = getNetwork();
+  private async syncAccount({
+    address,
+    isConnected,
+    chainId,
+    connector,
+    isConnecting,
+    isReconnecting
+  }: Pick<
+    GetAccountReturnType,
+    'address' | 'isConnected' | 'chainId' | 'connector' | 'isConnecting' | 'isReconnecting'
+  >) {
     this.resetAccount();
-    if (isConnected && address && chain) {
-      const caipAddress: CaipAddress = `${ConstantsUtil.EIP155}:${chain.id}:${address}`;
+    this.syncNetwork(address, chainId, isConnected);
+    this.setLoading(!!connector && (isConnecting || isReconnecting));
+
+    if (isConnected && address && chainId) {
+      const caipAddress: CaipAddress = `${ConstantsUtil.EIP155}:${chainId}:${address}`;
       this.setIsConnected(isConnected);
       this.setCaipAddress(caipAddress);
       await Promise.all([
-        this.syncProfile(address),
-        this.syncBalance(address, chain),
+        this.syncProfile(address, chainId),
+        this.syncBalance(address, chainId),
+        this.syncConnectedWalletInfo(connector),
         this.getApprovedCaipNetworksData()
       ]);
       this.hasSyncedConnectedAccount = true;
@@ -218,79 +236,117 @@ export class Web3Modal extends Web3ModalScaffold {
     }
   }
 
-  private async syncNetwork() {
-    const { address, isConnected } = getAccount();
-    const { chain } = getNetwork();
+  private async syncNetwork(address?: Hex, chainId?: number, isConnected?: boolean) {
+    const chain = this.wagmiConfig.chains.find((c: Chain) => c.id === chainId);
 
-    if (chain) {
-      const chainId = String(chain.id);
-      const caipChainId: CaipNetworkId = `${ConstantsUtil.EIP155}:${chainId}`;
+    if (chain || chainId) {
+      const name = chain?.name ?? chainId?.toString();
+      const id = Number(chain?.id ?? chainId);
+      const caipChainId: CaipNetworkId = `${ConstantsUtil.EIP155}:${id}`;
       this.setCaipNetwork({
         id: caipChainId,
-        name: chain.name,
-        imageId: PresetsUtil.EIP155NetworkImageIds[chain.id],
-        imageUrl: this.options?.chainImages?.[chain.id]
+        name,
+        imageId: PresetsUtil.EIP155NetworkImageIds[id],
+        imageUrl: this.options?.chainImages?.[id]
       });
-      if (isConnected && address) {
-        const caipAddress: CaipAddress = `${ConstantsUtil.EIP155}:${chain.id}:${address}`;
+      if (isConnected && address && chainId) {
+        const caipAddress: CaipAddress = `${ConstantsUtil.EIP155}:${id}:${address}`;
         this.setCaipAddress(caipAddress);
-        if (chain.blockExplorers?.default?.url) {
+        if (chain?.blockExplorers?.default?.url) {
           const url = `${chain.blockExplorers.default.url}/address/${address}`;
           this.setAddressExplorerUrl(url);
         } else {
           this.setAddressExplorerUrl(undefined);
         }
         if (this.hasSyncedConnectedAccount) {
-          await this.syncBalance(address, chain);
+          await this.syncProfile(address, chainId);
+          await this.syncBalance(address, chainId);
         }
       }
     }
   }
 
-  private async syncProfile(address: Address) {
+  private async syncProfile(address: Hex, chainId: number) {
     try {
-      const response = await this.fetchIdentity({
-        caipChainId: `${ConstantsUtil.EIP155}:${mainnet.id}`,
-        address
-      });
+      const response = await this.fetchIdentity({ address });
 
       if (!response) {
         throw new Error('Couldnt fetch idendity');
       }
 
-      const { name, avatar } = response;
-
-      this.setProfileName(name);
-      this.setProfileImage(avatar);
+      this.setProfileName(response.name);
+      this.setProfileImage(response.avatar);
     } catch {
-      const profileName = await fetchEnsName({ address, chainId: mainnet.id });
-      if (profileName) {
-        this.setProfileName(profileName);
-        const profileImage = await fetchEnsAvatar({ name: profileName, chainId: mainnet.id });
-        if (profileImage) {
-          this.setProfileImage(profileImage);
+      if (chainId === mainnet.id) {
+        const profileName = await getEnsName(this.wagmiConfig, { address, chainId });
+        if (profileName) {
+          this.setProfileName(profileName);
+          const profileImage = await wagmiGetEnsAvatar(this.wagmiConfig, {
+            name: profileName,
+            chainId
+          });
+          if (profileImage) {
+            this.setProfileImage(profileImage);
+          }
         }
+      } else {
+        this.setProfileName(undefined);
+        this.setProfileImage(undefined);
       }
     }
   }
 
-  private async syncBalance(address: Address, chain: Chain) {
-    const balance = await fetchBalance({
-      address,
-      chainId: chain.id,
-      token: this.options?.tokens?.[chain.id]?.address as Address
-    });
-    this.setBalance(balance.formatted, balance.symbol);
+  private async syncBalance(address: Hex, chainId: number) {
+    const chain = this.wagmiConfig.chains.find((c: Chain) => c.id === chainId);
+    if (chain) {
+      const balance = await getBalance(this.wagmiConfig, {
+        address,
+        chainId: chain.id
+      });
+      const formattedBalance = formatUnits(balance.value, balance.decimals);
+      this.setBalance(formattedBalance, balance.symbol);
+
+      return;
+    }
+    this.setBalance(undefined, undefined);
   }
 
-  private syncConnectors(wagmiConfig: Web3ModalClientOptions['wagmiConfig']) {
+  private async syncConnectedWalletInfo(connector: GetAccountReturnType['connector']) {
+    if (!connector) {
+      throw Error('syncConnectedWalletInfo - connector is undefined');
+    }
+
+    if (connector.id === ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID && connector.getProvider) {
+      const walletConnectProvider = (await connector.getProvider()) as Awaited<
+        ReturnType<(typeof EthereumProvider)['init']>
+      >;
+      if (walletConnectProvider.session) {
+        this.setConnectedWalletInfo({
+          ...walletConnectProvider.session.peer.metadata,
+          name: walletConnectProvider.session.peer.metadata.name,
+          icon: walletConnectProvider.session.peer.metadata.icons?.[0]
+        });
+      }
+    } else {
+      this.setConnectedWalletInfo({ name: connector.name, icon: connector.icon });
+    }
+  }
+
+  private syncConnectors(connectors: Web3ModalClientOptions['wagmiConfig']['connectors']) {
+    const uniqueIds = new Set();
+    const filteredConnectors = connectors.filter(
+      item => !uniqueIds.has(item.id) && uniqueIds.add(item.id)
+    );
+
+    const excludedConnectors = [ConstantsUtil.EMAIL_CONNECTOR_ID];
+
     const w3mConnectors: Connector[] = [];
-    wagmiConfig.connectors.forEach(({ id, name }) => {
-      if (![ConstantsUtil.EMAIL_CONNECTOR_ID].includes(id)) {
+    filteredConnectors.forEach(({ id, name, icon }) => {
+      if (!excludedConnectors.includes(id)) {
         w3mConnectors.push({
           id,
           explorerId: PresetsUtil.ConnectorExplorerIds[id],
-          imageId: PresetsUtil.ConnectorImageIds[id],
+          imageId: PresetsUtil.ConnectorImageIds[id] ?? icon,
           imageUrl: this.options?.connectorImages?.[id],
           name: PresetsUtil.ConnectorNamesMap[id] ?? name,
           type: PresetsUtil.ConnectorTypesMap[id] ?? 'EXTERNAL'
@@ -299,12 +355,13 @@ export class Web3Modal extends Web3ModalScaffold {
     });
 
     this.setConnectors(w3mConnectors);
+    this.syncEmailConnector(filteredConnectors);
   }
 
-  private async syncEmailConnector(wagmiConfig: Web3ModalClientOptions['wagmiConfig']) {
-    const emailConnector = wagmiConfig.connectors.find(
-      ({ id }) => id === ConstantsUtil.EMAIL_CONNECTOR_ID
-    );
+  private async syncEmailConnector(
+    connectors: Web3ModalClientOptions['wagmiConfig']['connectors']
+  ) {
+    const emailConnector = connectors.find(({ id }) => id === ConstantsUtil.EMAIL_CONNECTOR_ID);
     if (emailConnector) {
       const provider = await emailConnector.getProvider();
       this.addConnector({
@@ -316,8 +373,10 @@ export class Web3Modal extends Web3ModalScaffold {
     }
   }
 
-  private async listenEmailConnector(wagmiConfig: Web3ModalClientOptions['wagmiConfig']) {
-    const connector = wagmiConfig.connectors.find(c => c.id === ConstantsUtil.EMAIL_CONNECTOR_ID);
+  private async listenEmailConnector(
+    connectors: Web3ModalClientOptions['wagmiConfig']['connectors']
+  ) {
+    const connector = connectors.find(c => c.id === ConstantsUtil.EMAIL_CONNECTOR_ID);
 
     const connectedConnector = await StorageUtil.getItem('@w3m/connected_connector');
     if (connector && connectedConnector === 'EMAIL') {