+++ apps/native-cli/CHANGELOG.md
@@ -1,11 +1,25 @@
 # @apps/native-cli
 
-## 1.0.7
+## 1.0.7-alpha.2
 
 ### Patch Changes
 
 - Updated dependencies
-  - @web3modal/wagmi-react-native@1.4.1
+  - @web3modal/wagmi-react-native@2.0.0-alpha.2
+
+## 1.0.7-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @web3modal/wagmi-react-native@2.0.0-alpha.1
+
+## 1.0.7-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @web3modal/wagmi-react-native@2.0.0-alpha.0
 
 ## 1.0.6