+++ packages/ethers5/CHANGELOG.md
@@ -1,13 +1,34 @@
 # @web3modal/ethers5-react-native
 
-## 1.4.1
+## 2.0.0-alpha.2
 
 ### Patch Changes
 
-- email wallet fixes
+- alpha.2
 - Updated dependencies
-  - @web3modal/scaffold-react-native@1.4.1
-  - @web3modal/scaffold-utils-react-native@1.4.1
+  - @web3modal/scaffold-utils-react-native@2.0.0-alpha.2
+  - @web3modal/scaffold-react-native@2.0.0-alpha.2
+
+## 2.0.0-alpha.1
+
+### Patch Changes
+
+- fixes
+- Updated dependencies
+  - @web3modal/scaffold-utils-react-native@2.0.0-alpha.1
+  - @web3modal/scaffold-react-native@2.0.0-alpha.1
+
+## 2.0.0-alpha.0
+
+### Major Changes
+
+- 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @web3modal/scaffold-react-native@2.0.0-alpha.0
+  - @web3modal/scaffold-utils-react-native@2.0.0-alpha.0
 
 ## 1.4.0